refactor: 패키지 구조 보완
