+++ src/main/java/racingcar/view/inputview/InputDevice.java
@@ -1,4 +1,4 @@
-package racingcar.inputview;
+package racingcar.view.inputview;
 
 public interface InputDevice {
     int getIntNumber();
