+++ src/main/java/racingcar/view/inputview/InputHandler.java
@@ -1,4 +1,4 @@
-package racingcar.inputview;
+package racingcar.view.inputview;
 
 import racingcar.dto.Input;
