+++ src/main/java/racingcar/view/resultview/ConsoleResultView.java
@@ -1,4 +1,4 @@
-package racingcar.resultview;
+package racingcar.view.resultview;
 
 import racingcar.domain.RacingCar;
