+++ src/test/java/racingcar/InputHandlerTest.java
@@ -4,9 +4,9 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import racingcar.dto.Input;
-import racingcar.inputview.ConsoleInputView;
-import racingcar.inputview.InputHandler;
 import racingcar.mock.MockInputDevice;
+import racingcar.view.inputview.ConsoleInputView;
+import racingcar.view.inputview.InputHandler;
 
 import static org.assertj.core.api.Assertions.assertThat;
